pyproject.toml
[tool.poetry]
name = "healthdbETL"
version = "0.1.0"
description = "ETL and Database Interaction Tools for HealthDB"
authors = ["Joseph Deferio <jdeferio.io@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
FHIR-Parser = "^0.1.5"
pandas = "^1.2.0"
openpyxl = "^3.0.5"
SQLAlchemy = "^1.3.22"
boto3 = "^1.16.54"
numpy = "^1.19.5"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
isort = "^5.7.0"
docformatter = "^1.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
